Move InstallRequirement.from_line to constructors module
