+++ src/pip/_internal/operations/freeze.py
@@ -10,8 +10,9 @@ from pip._vendor.packaging.utils import canonicalize_name
 from pip._vendor.pkg_resources import RequirementParseError
 
 from pip._internal.exceptions import InstallationError
-from pip._internal.req import InstallRequirement
-from pip._internal.req.constructors import install_req_from_editable
+from pip._internal.req.constructors import (
+    install_req_from_editable, install_req_from_line,
+)
 from pip._internal.req.req_file import COMMENT_RE
 from pip._internal.utils.deprecation import deprecated
 from pip._internal.utils.misc import (
@@ -106,7 +107,7 @@ def freeze(
                             wheel_cache=wheel_cache,
                         )
                     else:
-                        line_req = InstallRequirement.from_line(
+                        line_req = install_req_from_line(
                             COMMENT_RE.sub('', line).strip(),
                             isolated=isolated,
                             wheel_cache=wheel_cache,
